Ewon/Docker/requirements.txt
Adafruit-Blinka==8.18.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-motor==3.4.9
adafruit-circuitpython-pca9685==3.4.8
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.13.2
adafruit-circuitpython-servokit==1.3.13
adafruit-circuitpython-typing==1.9.2
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
azure-cognitiveservices-speech==1.28.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.1
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nextion==1.8.1
notebook==6.5.4
notebook-shim==0.2.3
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyftdi==0.54.0
Pygments==2.15.1
pyrsistent==0.19.3
pyserial==3.5
pyserial-asyncio==0.6
python-dateutil==2.8.2
python-json-logger==2.0.7
pyusb==1.2.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.29.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sysv-ipc==1.1.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
typing-extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
yarg==0.1.9
zipp==3.15.0
Flask==2.1.0